Extract smaller component
